+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 * Renamed `override` to `bypass` in the engineering mode keywords and commands to match the labelling in the PLC.
 * Output whether the engineering bypass is local or remote.
-
+* Changed the default engineering mode duration to 12 hours.
 
 ## 1.2.2 - January 24, 2025